re ZON-3213: Move tests from `test_parse` to `test_mobile` and adjust them.
